fix(ci): resolve LOC violations in quality gate and split oversized source files
- Fix quality_gate.sh exclusion patterns (add ./ prefix for find output)
- Add missing exclusions: tests/, .agents/, .claude/, .opencode/, plans/, public/
- Split 9 source files under 600-line limit using barrel re-exports
- All modules preserve existing public API (zero import changes)
- LOC check: 0 violations (was 40+)

Files split:
- worker/lib/research-agent/fetcher.ts (1106→100 lines)
- worker/lib/d1/queries.ts (974→36 lines)
- worker/lib/d1/migrations.ts (799→15 lines)
- worker/lib/validation/reward-scraper.ts (763→30 lines)
- worker/lib/validation/url-validator.ts (738→486 lines)
- worker/lib/validation/code-validator.ts (719→306 lines)
- worker/lib/referral-storage/dual-write.ts (651→416 lines)
- worker/types.ts (681→6 lines)
- worker/lib/research-agent/types.ts (640→182 lines)
